+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## `@svizzle/barchart` v0.5.0 (next)
 
 - add prop `titleFontSize`
+- add prop `selectedKeys`
 - don't show the axis with positive and zeroes or negative and zeroes
 
 ## `@svizzle/choropleth` v0.4.0 (next)
@@ -31,6 +32,11 @@
 ### Array -> Array
 
 - add `setIndexAsKey`
+
+### Array -> Object
+
+- add `makeKeyedFalse`
+- add `makeKeyedTrue`
 
 ### String -> Number